BLD: setup Webpack and Babel for ES6
