+++ .babelrc
@@ -1,0 +1,7 @@
+{
+    "presets": ["@babel/preset-env"],
+    "plugins": [
+        "@babel/plugin-proposal-class-properties",
+        "@babel/plugin-transform-exponentiation-operator"
+    ]
+}
