+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode
 
 .notupload
+node_modules
